Clean up and refactor
Removing GET and DELETE from base class, removing obsolete methods, using prepareRequest(), ignoring invalid tests, using order builder
